+++ src/main/java/shipeditor/utility/components/dialog/DialogUtilities.java
@@ -65,4 +65,9 @@ public final class DialogUtilities {
         return null;
     }
 
+    public static void showSlotCreationDialog() {
+        SlotCreationDialog dialog = new SlotCreationDialog(shipeditor.PrimaryWindow.getInstance());
+        dialog.setVisible(true);
+    }
+
 }